+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [UNRELEASED]
 
 - added support for message queue while task is processing
+- making sure to include user message when using Gemini models to avoid signature error
 
 ## [0.53.0]